tracing fix

+++ Xtricate.DocSet/DocStorage.cs
@@ -34,10 +34,13 @@ namespace Xtricate.DocSet
             TableName = options.GetTableName<TDoc>();
             IndexMaps = indexMap.NullToEmpty().ToList().Where(im => im != null).OrderBy(i => i.Name);
 
-            Trace.WriteLine($"table: {TableName}");
-            IndexMaps.ForEach(im => Trace.WriteLine($"index map: {typeof (TDoc).Name} > {im.Name} [{im.Description}]"));
-            Trace.WriteLine($"connection: {Options.ConnectionString}");
-
+            if (Options.EnableTracing)
+            {
+                Trace.WriteLine($"table: {TableName}");
+                IndexMaps.ForEach(
+                    im => Trace.WriteLine($"index map: {typeof (TDoc).Name} > {im.Name} [{im.Description}]"));
+                Trace.WriteLine($"connection: {Options.ConnectionString}");
+            }
             Initialize();
         }
 
@@ -85,7 +88,8 @@ namespace Xtricate.DocSet
                 StorageAction result;
                 if (!forceInsert && Exists(key, tags))
                 {
-                    Trace.WriteLine($"{TableName} update: key={key},tags={tags?.ToString("||")}");
+                    if (Options.EnableTracing)
+                        Trace.WriteLine($"{TableName} update: key={key},tags={tags?.ToString("||")}");
                     sql =
                         $@"
     UPDATE {TableName
@@ -104,7 +108,8 @@ namespace Xtricate.DocSet
                 }
                 else
                 {
-                    Trace.WriteLine($"{TableName} insert: key={key},tags={tags?.ToString("||")}");
+                    if (Options.EnableTracing)
+                        Trace.WriteLine($"{TableName} insert: key={key},tags={tags?.ToString("||")}");
                     sql =
                         $@"
     INSERT INTO {TableName
@@ -136,7 +141,8 @@ namespace Xtricate.DocSet
 
         public virtual long Count(IEnumerable<string> tags = null, IEnumerable<Criteria> criterias = null)
         {
-            Trace.WriteLine($"{TableName} count: tags={tags?.ToString("||")}");
+            if (Options.EnableTracing)
+                Trace.WriteLine($"{TableName} count: tags={tags?.ToString("||")}");
 
             using (var conn = CreateConnection())
             {
@@ -153,7 +159,8 @@ namespace Xtricate.DocSet
             DateTime? fromDateTime = null, DateTime? tillDateTime = null,
             int skip = 0, int take = 0)
         {
-            Trace.WriteLine(
+            if (Options.EnableTracing)
+                Trace.WriteLine(
                 $"{TableName} load: key={key}, tags={tags?.ToString("||")}, criterias={criterias?.Select(c => c.Name + ":" + c.Value).ToString("||")}");
 
             using (var conn = CreateConnection())
@@ -176,7 +183,8 @@ namespace Xtricate.DocSet
             DateTime? fromDateTime = null, DateTime? tillDateTime = null,
             int skip = 0, int take = 0)
         {
-            Trace.WriteLine(
+            if (Options.EnableTracing)
+                Trace.WriteLine(
                 $"{TableName} load: tags={tags?.ToString("||")}, criterias={criterias?.Select(c => c.Name + ":" + c.Value).ToString("||")}");
 
             using (var conn = CreateConnection())
@@ -197,7 +205,8 @@ namespace Xtricate.DocSet
         public virtual StorageAction Delete(object key, IEnumerable<string> tags = null,
             IEnumerable<Criteria> criterias = null)
         {
-            Trace.WriteLine($"{TableName} delete: key={key},tags={tags?.ToString("||")}");
+            if (Options.EnableTracing)
+                Trace.WriteLine($"{TableName} delete: key={key},tags={tags?.ToString("||")}");
             using (var conn = CreateConnection())
             {
                 var sql = $@"DELETE FROM {TableName} WHERE [key]='{key}'";
@@ -212,7 +221,8 @@ namespace Xtricate.DocSet
         public virtual StorageAction Delete(IEnumerable<string> tags,
             IEnumerable<Criteria> criterias = null)
         {
-            Trace.WriteLine($"{TableName} delete: tags={tags?.ToString("||")}");
+            if (Options.EnableTracing)
+                Trace.WriteLine($"{TableName} delete: tags={tags?.ToString("||")}");
             if (tags.IsNullOrEmpty()) return StorageAction.None;
             using (var conn = CreateConnection())
             {
@@ -259,7 +269,8 @@ namespace Xtricate.DocSet
             using (var conn = CreateConnection())
             {
                 conn.Open();
-                Trace.WriteLine($"{tableName} exists [{conn.Database}]");
+                if (Options.EnableTracing)
+                    Trace.WriteLine($"{tableName} exists [{conn.Database}]");
                 return
                     conn.Query<string>(SqlBuilder.TableNamesSelect())
                         .Any(t => t.Equals(tableName, StringComparison.InvariantCultureIgnoreCase));
@@ -272,7 +283,8 @@ namespace Xtricate.DocSet
             using (var conn = CreateConnection())
             {
                 conn.Open();
-                Trace.WriteLine($"{options.SchemaName} ensure schema [{conn.Database}]");
+                if (Options.EnableTracing)
+                    Trace.WriteLine($"{options.SchemaName} ensure schema [{conn.Database}]");
                 if (conn.Query<string>(@"
     SELECT QUOTENAME(TABLE_SCHEMA) AS Name
     FROM INFORMATION_SCHEMA.TABLES")
@@ -286,7 +298,9 @@ namespace Xtricate.DocSet
                 }
                 catch (SqlException e)
                 {
-                    Trace.WriteLine($"create schema: {e.Message}: ");
+                    // swallog
+                    if (Options.EnableTracing)
+                        Trace.WriteLine($"create schema: {e.Message}: ");
                 }
             }
         }
@@ -297,7 +311,8 @@ namespace Xtricate.DocSet
             using (var conn = CreateConnection())
             {
                 conn.Open();
-                Trace.WriteLine($"{tableName} ensure table [{conn.Database}]");
+                if (Options.EnableTracing)
+                    Trace.WriteLine($"{tableName} ensure table [{conn.Database}]");
                 // http://stackoverflow.com/questions/11938044/what-are-the-best-practices-for-using-a-guid-as-a-primary-key-specifically-rega
                 var sql = string.Format(@"
     CREATE TABLE {0}(
@@ -325,7 +340,8 @@ namespace Xtricate.DocSet
             using (var conn = CreateConnection())
             {
                 conn.Open();
-                Trace.WriteLine(
+                if (Options.EnableTracing)
+                    Trace.WriteLine(
                     $"{tableName} ensure index [{conn.Database}], index={IndexMaps.NullToEmpty().Select(i => i.Name).ToString(", ")}");
                 var sql = IndexMaps.NullToEmpty().Select(i =>
                     string.Format(@"
@@ -347,7 +363,8 @@ namespace Xtricate.DocSet
             using (var conn = CreateConnection())
             {
                 conn.Open();
-                Trace.WriteLine($"{tableName} drop table [{conn.Database}]");
+                if (Options.EnableTracing)
+                    Trace.WriteLine($"{tableName} drop table [{conn.Database}]");
                 var sql = string.Format(@"DROP TABLE {0}", tableName);
                 conn.Execute(sql);
             }
@@ -359,7 +376,8 @@ namespace Xtricate.DocSet
             using (var conn = CreateConnection())
             {
                 conn.Open();
-                Trace.WriteLine($"{tableName} drop table [{conn.Database}]");
+                if (Options.EnableTracing)
+                    Trace.WriteLine($"{tableName} drop table [{conn.Database}]");
                 var sql = IndexMaps.NullToEmpty().Select(i =>
                     string.Format(@"
     IF EXISTS(SELECT * FROM sys.columns
